Refactor menu toggle functionality in Menu and NavBar components
